making the collapsible map section component agnostic towards which map component is used. Simplifying the MapSVG component into one file.
